+++ src/App/Deploy/Laravel.php
@@ -2,27 +2,19 @@
 
 namespace Console\App\Deploy;
 
-use Art4\JsonApiClient\Helper\Parser;
-use Art4\JsonApiClient\V1\Document;
-use Console\App\Commands\Command;
-use Console\App\Commands\Databases\DatabasesDescribeCommand;
 use Console\App\Commands\Files\FilesDeleteCommand;
 use Console\App\Commands\Files\FilesUpdateCommand;
-use Console\App\Commands\Files\FilesUploadCommand;
-use Console\App\Helpers\DeployHelper;
 use Dotenv\Dotenv;
 use Exception;
 use GuzzleHttp\Exception\ClientException;
-use Symfony\Component\Console\Input\ArrayInput;
-use Symfony\Component\Console\Output\BufferedOutput;
 use GuzzleHttp\Exception\GuzzleException;
 
 class Laravel extends DeployAbstract
 {
-	/**
-	 * @var array
-	 */
-	private $localEnv = [];
+
+	const SKIP_COMMANDS = [
+		'artisan migrate',
+	];
 
 	/**
 	 * @param string $appPath
@@ -35,11 +27,9 @@ class Laravel extends DeployAbstract
 	{
 		parent::deployApp($appPath, $isFirstDeploy);
 		(Dotenv::create($this->appPath))->load();
-		$this->localEnv = $_ENV;
-		$this->updateEnvFileToUpload();
+		$this->updateEnvFileToUpload($_ENV, $this->prepareEnvFile($_ENV));
 		$zip = $this->getZipApp();
-
-		$this->restoreLocalEnvFile();
+		$this->restoreLocalEnvFile($_ENV);
 		$this->uploadToApp($zip, $this->releaseFolder . self::ARCHIVE_NAME);
 		$this->unarchiveApp($this->releaseFolder . self::ARCHIVE_NAME);
 		$this->deleteArchiveRemote($this->releaseFolder . self::ARCHIVE_NAME);
@@ -47,50 +37,26 @@ class Laravel extends DeployAbstract
 		$this->setUpPermissions();
 		if ($this->isFirstDeploy) {
 			if ($this->config['database']['type'] == 'external') {
-				$this->createDatabase();
+				$this->createDatabase(getenv('DB_DATABASE'));
 				if (!empty($this->config['database']['sql_dump'])) {
-					$this->importSqlDump();
+					$this->importSqlDump(getenv('DB_DATABASE'));
 				}
 			} elseif ($this->config['database']['system'] == 'sqlite') {
 				$this->initSqliteDatabase();
 			} else {
 				$this->initDatabase();
 				$this->createDatabaseUser();
-				$this->createDatabase();
+				$this->createDatabase(getenv('DB_DATABASE'));
 				if (!empty($this->config['database']['sql_dump'])) {
-					$this->importSqlDump();
+					$this->importSqlDump(getenv('DB_DATABASE'));
 				}
 			}
 		}
 		$dbBackupId = $this->backupDatabase();
-		$this->runCommands();
+		$this->runCommands(self::SKIP_COMMANDS);
 		$this->artisanMigrate($dbBackupId);
 		$this->symlinkRelease($this->releaseFolder . 'public', 'Linking your current release', $this->isFirstDeploy);
 		$this->deleteArchiveLocal();
-	}
-
-	/**
-	 * @throws Exception
-	 */
-	private function runCommands()
-	{
-		$step = 'runCommands';
-		$this->setStep($step, function () {
-			return;
-		});
-		if (!empty($this->config['commands'])) {
-			foreach ($this->config['commands'] as $command) {
-				if (preg_match('/artisan migrate/', $command)) {
-					continue;
-				}
-				$this->appRunCommand(
-					$this->config['app']['id'],
-					'cd ' . $this->releaseFolder . ' && ' . $command,
-					$command
-				);
-			}
-		}
-		$this->updateStepToSuccess($step);
 	}
 
 	/**
@@ -201,32 +167,6 @@ class Laravel extends DeployAbstract
 	/**
 	 *
 	 */
-	private function restoreLocalEnvFile()
-	{
-		$step = 'restoreLocalEnvFile';
-		$this->setStep($step, function () {
-			return;
-		});
-		$this->updateEnvFile($this->localEnv);
-		$this->updateStepToSuccess($step);
-	}
-
-	/**
-	 *
-	 */
-	private function updateEnvFileToUpload()
-	{
-		$step = 'updateEnvFileToUpload';
-		$this->setStep($step, function () {
-			$this->updateEnvFile($this->localEnv);
-		});
-		$this->updateEnvFile($this->prepareEnvFile());
-		$this->updateStepToSuccess($step);
-	}
-
-	/**
-	 *
-	 */
 	public function revertProcess()
 	{
 		$this->consoleOutput->writeln('<comment>Starting revert</comment>');
@@ -238,9 +178,10 @@ class Laravel extends DeployAbstract
 	}
 
 	/**
+	 * @param array $localEnv
 	 * @return array
 	 */
-	private function prepareEnvFile(): array
+	private function prepareEnvFile(array $localEnv): array
 	{
 		if ($this->config['database']['system'] == 'sqlite') {
 			$env = [
@@ -258,19 +199,8 @@ class Laravel extends DeployAbstract
 			];
 		}
 		$envFromConfig = !empty($this->config['environment']) ? $this->config['environment'] : [];
-		$newEnv = array_merge($envFromConfig, $env);
-		return array_merge($this->localEnv, $newEnv);
-	}
-
-	/**
-	 * @param array $newEnvVars
-	 */
-	private function updateEnvFile(array $newEnvVars)
-	{
-		file_put_contents($this->appPath . '.env', '');
-		foreach ($newEnvVars as $key => $val) {
-			file_put_contents($this->appPath . '.env', $key . '=' . $val . PHP_EOL, FILE_APPEND);
-		}
+		$remoteEnv = array_merge($envFromConfig, $env);
+		return array_merge($localEnv, $remoteEnv);
 	}
 
 	/**
@@ -290,249 +220,6 @@ class Laravel extends DeployAbstract
 			'php ' . $this->releaseFolder . 'artisan migrate --force',
 			'Migrating schema'
 		);
-		$this->updateStepToSuccess($step);
-	}
-
-	/**
-	 * @throws Exception
-	 */
-	private function createDatabaseUser()
-	{
-		$step = 'createDatabaseUser';
-		$this->setStep($step, function () {
-			$command = sprintf(
-				'mysql --user=root --host=%s --password=%s --execute "DROP USER \'%s\'"',
-				$this->config['database']['connection']['host'],
-				$this->config['database']['root_password'],
-				$this->config['database']['connection']['user']
-			);
-			$this->appRunCommand(
-				$this->config['app']['id'],
-				$command,
-				'Drop database user ' . $this->config['database']['connection']['user']
-			);
-		});
-		$command = sprintf(
-			'mysql --user=root --host=%s --password=%s --execute "CREATE USER \'%s\'@\'%%\' IDENTIFIED WITH mysql_native_password BY \'%s\';GRANT ALL PRIVILEGES ON * . * TO \'%s\'@\'%%\';FLUSH PRIVILEGES;"',
-			$this->config['database']['connection']['host'],
-			$this->config['database']['root_password'],
-			$this->config['database']['connection']['user'],
-			$this->config['database']['connection']['password'],
-			$this->config['database']['connection']['user']
-		);
-		$this->appRunCommand(
-			$this->config['app']['id'],
-			$command,
-			'Creating database user ' . $this->config['database']['connection']['user']
-		);
-		$this->updateStepToSuccess($step);
-	}
-
-	/**
-	 * @throws Exception
-	 */
-	private function initDatabase()
-	{
-		$progressBar = Command::getProgressBar('Initializing database', $this->consoleOutput);
-		$progressBar->start();
-		$dbIsRunning = false;
-		while (!$dbIsRunning) {
-			$dbIsRunning = $this->isDbAlreadyRunning($this->config['database']['connection']['host']);
-			$progressBar->advance();
-		}
-		$counter = 0;
-		/** We need this hack with sleep, because status of database is running,
-		 * but it still not ready for connections so ~30-40 secs need to wait
-		 */
-		while ($counter != 50) {
-			$progressBar->advance();
-			$counter++;
-			sleep(1);
-		}
-		$progressBar->finish();
-		$this->consoleOutput->write(PHP_EOL);
-	}
-
-	/**
-	 * @throws Exception
-	 */
-	private function createDatabase()
-	{
-		$step = 'CreateDatabase';
-		$this->setStep($step, function () {
-			$command = sprintf(
-				'mysql --user=%s --host=%s --password=%s --execute "drop database %s;"',
-				$this->config['database']['connection']['user'],
-				$this->config['database']['connection']['host'],
-				$this->config['database']['connection']['password'],
-				getenv('DB_DATABASE')
-			);
-			$this->appRunCommand(
-				$this->config['app']['id'],
-				$command,
-				'Drop database'
-			);
-		});
-
-		$command = sprintf(
-			'mysql --user=%s --host=%s --password=%s --execute "create database %s;"',
-			$this->config['database']['connection']['user'],
-			$this->config['database']['connection']['host'],
-			$this->config['database']['connection']['password'],
-			getenv('DB_DATABASE')
-		);
-		$this->appRunCommand(
-			$this->config['app']['id'],
-			$command,
-			'Creating database `' . getenv('DB_DATABASE') . '`'
-		);
-		$this->updateStepToSuccess($step);
-	}
-
-	/**
-	 * @throws Exception
-	 * @throws GuzzleException
-	 */
-	private function initSqliteDatabase()
-	{
-		$step = 'initSqliteDatabase';
-		$this->setStep($step, function () {
-			return;
-		});
-		if (!empty($this->config['database']['sql_dump'])) {
-			$filesUploadCommand = $this->application->find(FilesUploadCommand::getDefaultName());
-			$args = [
-				'command'     => FilesUploadCommand::getDefaultName(),
-				'file'        => $this->config['database']['sql_dump'],
-				'app_id'      => $this->config['app']['id'],
-				'remote_path' => DeployHelper::SQLITE_RELATIVE_REMOTE_PATH,
-				'--json'      => true,
-			];
-			if ($filesUploadCommand->run(new ArrayInput($args), $this->consoleOutput) != '0') {
-				throw new Exception('Cant upload mysqli dump');
-			}
-			$this->consoleOutput->write(PHP_EOL);
-		} else {
-			$url = sprintf(
-				FilesUploadCommand::API_ENDPOINT,
-				$this->config['app']['id']
-			);
-			$this->sendRequest($url, 'POST', 'Creating sqlite database', json_encode([
-				'data' => [
-					'attributes' => [
-						'apache_writable' => true,
-					],
-					'id'         => DeployHelper::SQLITE_RELATIVE_REMOTE_PATH,
-					'type'       => 'files',
-				],
-			]));
-		}
-		$this->updateStepToSuccess($step);
-	}
-
-	/**
-	 * @throws Exception
-	 */
-	private function importSqlDump()
-	{
-		$step = 'ImportSqlDump';
-		$this->setStep($step, function () {
-			return;
-		});
-		$filesUploadCommand = $this->application->find(FilesUploadCommand::getDefaultName());
-		$args = [
-			'command'     => FilesUploadCommand::getDefaultName(),
-			'file'        => $this->config['database']['sql_dump'],
-			'app_id'      => $this->config['app']['id'],
-			'remote_path' => $this->releaseFolder . self::SQL_DUMP_NAME,
-			'--json'      => true,
-		];
-		if ($filesUploadCommand->run(new ArrayInput($args), $this->consoleOutput) == '0') {
-			$command = sprintf(
-				'mysql -u %s --host=%s --password=%s  %s < %s',
-				$this->config['database']['connection']['user'],
-				$this->config['database']['connection']['host'],
-				$this->config['database']['connection']['password'],
-				getenv('DB_DATABASE'),
-				$this->releaseFolder . self::SQL_DUMP_NAME
-			);
-			$this->appRunCommand(
-				$this->config['app']['id'],
-				$command,
-				'Importing sql dump'
-			);
-			$this->updateStepToSuccess($step);
-		} else {
-			throw new Exception('Uploading sql dump to app failed');
-		}
-	}
-
-
-	/**
-	 * @param string $dbId
-	 * @return bool
-	 * @throws Exception
-	 */
-	protected function isDbAlreadyRunning(string $dbId): bool
-	{
-		$appRunsNewCommand = $this->application->find(DatabasesDescribeCommand::getDefaultName());
-		$args = [
-			'command'     => DatabasesDescribeCommand::getDefaultName(),
-			'database_id' => $dbId,
-			'--json'      => true,
-		];
-		$bufferOutput = new BufferedOutput();
-		if ($appRunsNewCommand->run(new ArrayInput($args), $bufferOutput) == '0') {
-			/** @var Document $document */
-			$document = Parser::parseResponseString($bufferOutput->fetch());
-			return $document->get('data.attributes.status') === 'running';
-		} else {
-			throw new Exception('Checking db status failed');
-		}
-	}
-
-	/**
-	 * @param array $defaultDirs
-	 * @param array $skip
-	 * @param bool $recur
-	 * @throws GuzzleException
-	 */
-	protected function setUpPermissions(array $defaultDirs = [], array $skip = [], bool $recur = false)
-	{
-		/** TODO rewrite this method */
-		$step = 'setDirectoryPermissions';
-		$this->setStep($step, function () {
-			return;
-		});
-		if (empty($this->config['apache_permissions_dir'])) {
-			return;
-		}
-		$directories = array_map(function ($val) {
-			if (strpos($val, 'storage') === false) {
-				return $this->releaseFolder . $val;
-			}
-		}, $this->config['apache_permissions_dir']);
-
-		foreach ($directories as $directory) {
-			try {
-				$this->sendRequest(
-					sprintf(FilesUpdateCommand::API_ENDPOINT, $this->config['app']['id'], ''),
-					'PATCH',
-					'Setting apache writable ' . $directory,
-					json_encode([
-						'data' => [
-							'attributes' => [
-								'apache_writable' => true,
-							],
-							'id'         => $directory,
-							'type'       => 'files',
-						],
-					])
-				);
-			} catch (ClientException $clientException) {
-				$this->consoleOutput->writeln(PHP_EOL . '<comment>Directory ' . $directory . '/ not exists</comment>');
-			}
-		}
 		$this->updateStepToSuccess($step);
 	}
